pyproject.toml
[tool.poetry]
name = "dream"
version = "0.1.0"
description = ""
authors = ["vladislavkovalskyi <v.darknesssb@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mubble = "^1.0.0"
tortoise-orm = "^0.20.0"
envparse = "^0.2.0"
asyncpg = "^0.29.0"
openai = "^1.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
